Rename monitoring => status
