+++ tests/functional/api/view_status_test.py
@@ -1,8 +1,3 @@
-import pytest
-
-from tests.conftest import assert_cache_control
-
-
 class TestViewMonitoring:
     def test_status_view(self, test_app):
         response = test_app.get("/_status")
